# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/teamMembers": {
            "get": {
                "tags": [
                    "TeamMember"
                ],
                "summary": "getTeamMemberList",
                "description": "Get all TeamMembers",
                "operationId": "50e42d2216f357cb3fc2c9d04e56d568",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "TeamMember"
                ],
                "summary": "createTeamMember",
                "description": "Create TeamMember",
                "operationId": "793a2f9812ee40dbbf2e79904985274c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/teamMembers/{id}": {
            "get": {
                "tags": [
                    "TeamMember"
                ],
                "summary": "getTeamMemberItem",
                "description": "Get TeamMember",
                "operationId": "504f1e2454344367b03491dac070cd56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of TeamMember",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "TeamMember"
                ],
                "summary": "updateTeamMember",
                "description": "Update TeamMember",
                "operationId": "d17af343b44e4ce310e0f9288ec00d3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of TeamMember",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TeamMember"
                ],
                "summary": "deleteTeamMember",
                "description": "Delete TeamMember",
                "operationId": "87f4ab442a55d282e1ebade1a87da9c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of TeamMember",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}